move UPat and PatternMatcher from uopgraph.py to uops.py
towards instant UOps rewrite on UOp.alu

[run_process_replay]
